pyproject.toml
[project]
name = "langgraph-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.45,<0.3.0",
    "langchain-openai>=0.2.6,<0.3.0",
    "langserve>=0.3.0",
    "python-dotenv>=1.0.1",
    "fastapi[standard]>=0.115.4",
    "sse-starlette>=2.1.3",
    "langchain>=0.3.7,<0.4.0",
    "langchain-community>=0.2.11",
    "tavily-python>=0.5.0",
]
